+++ index.js
@@ -1,16 +1,13 @@
 /* eslint-disable no-unused-vars */
 /* eslint-disable no-restricted-globals */
 
+import {
+  container, contact, displayContact, displayBooks, displayForm, date, bookText,
+} from './modules/new.js';
+
 const form = document.querySelector('#form');
 const Author = document.querySelector('.Author');
 const Title = document.querySelector('.Title');
-const bookText = document.getElementById('book-text');
-const container = document.querySelector('.container');
-const contact = document.querySelector('#contact');
-const displayContact = document.querySelector('#display_contact');
-const displayBooks = document.querySelector('#display_list');
-const displayForm = document.querySelector('#display_form');
-const date = document.querySelector('.date');
 
 class Book {
   static validateInputData = (title, author) => {
@@ -93,55 +90,3 @@ form.addEventListener('submit', (e) => {
 window.deleteBook = (text1, text2) => {
   Book.removeBook(text1, text2);
 };
-
-contact.innerHTML = `
-        <h2> Contact information </h2>
-
-        <ul>
-          <li>Email Address: Kennediowusu@gmail.com</li>
-          <li>Phone Number: +233546985015</li>
-          <li>Street Address: Kwame Nkrumah Downside</li>
-        </ul>`;
-
-const showOnlyContact = () => {
-  contact.style.display = 'block';
-  container.style.display = 'none';
-  bookText.style.display = 'none';
-};
-
-const showBooksOnly = () => {
-  container.style.display = 'none';
-  bookText.style.display = 'block';
-  contact.style.display = 'none';
-  date.style.display = '';
-};
-
-const showOnlyForm = () => {
-  container.style.display = 'block';
-  bookText.style.display = 'none';
-  contact.style.display = 'none';
-  date.style.display = 'none';
-};
-
-function init() {
-  if (location.hash === '/#book-list') {
-    showBooksOnly();
-  } else if (location.hash === '/#form') {
-    showOnlyForm();
-  } else if (location.hash === '/#contact') {
-    showOnlyContact();
-  } else {
-    showOnlyForm();
-  }
-}
-
-window.document.addEventListener('DOMContentLoaded', init());
-displayContact.addEventListener('click', showOnlyContact);
-displayBooks.addEventListener('click', showBooksOnly);
-displayForm.addEventListener('click', showOnlyForm);
-
-function showDate() {
-  date.innerHTML = `${new Date()}`;
-}
-
-setInterval(showDate, 1000);
