+++ modules/retrieve_date.js
@@ -1,0 +1,7 @@
+import getDateAndTime from './date.js';
+
+const date = document.querySelector('.date');
+
+date.innerText = getDateAndTime();
+
+console.log(date.innerText);
